requirements.txt
Django==2.1.2
dj-ango==0.2.0
dj-database-url==0.5.0
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==3.0.1
psycopg2==2.7.5
pytz==2018.5
django-shell-plus==1.1.5
ipython==1.2.1
psycopg2-binary==2.7.5
gunicorn==19.6.0
django-heroku==0.3.1
whitenoise==4.1